+++ frontend/src/components/PLAYGROUND/Pages/ProjectsSandbox/index.tsx
@@ -1,5 +1,0 @@
-import { Table } from "components/PLAYGROUND/Table";
-
-export const ProjectsSandbox = () => {
-  return <Table apiEndPoint={"projects"} />;
-};