+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## Developer Changelog
 **May 15, 2023**
 - Frontend
+  - fixed issue with projects table not rendering
+
+**May 15, 2023**
+- Frontend
   - Upgraded to react-router v6.11
 
 **May 12, 2023 (DESCW-1102)**